Pringles can and refactored to a class
There are two versions of the Pringles Can engima machine that I can
find on the internet:

Cyber.org: https://cyber.org/find-curricula/pringles-can-enigma
Franklin Heath: http://wiki.franklinheath.co.uk/index.php/Enigma/Paper_Enigma

The latter stays faithful to the original Enigma configuration, but the
former simplifies things a little. This is primarily in the rotor
design, where the connections are nearer to each other to make it easier
to follow, but there are also a couple of differences in the
configuration and stepping of the rotors.

Configuration is done by aligning letters on the input ring, the rotors,
and the reflector. This means that the reflector can rotate, which
wasn't part of the original design.

Then the rotor stepping occurs after 26 characters, not when a certain
letter turns. So, in effect, the notch can be configured as well as the
ring setting.

Reading the instructions carefully, they also step from the wrong end,
and there's no mention of double-stepping.

These are reasonable simplifications, but did make it a pain to code!

Having coded that, I refactored the main code into a class.